+++ script/build-module.js
@@ -16,9 +16,9 @@ const GithubSlugger = require('github-slugger')
 const getIds = require('get-crowdin-file-ids')
 const remark = require('remark')
 const links = require('remark-inline-links')
-const visit = require('unist-util-visit')
 const parseElectronGlossary = require('../lib/parse-electron-glossary')
 const plaintextFix = require('../lib/remark-plaintext-fix')
+const fiddleUrls = require('../lib/remark-fiddle-urls')
 
 const contentDir = path.join(__dirname, '../content')
 const cheerio = require('cheerio')
@@ -29,25 +29,6 @@ const categoryNames = {
   tutorial: 'Guides',
 }
 const IGNORE_PATTERN = '<!-- i18n-ignore -->'
-
-// remark transformer for 'code' blocks to
-// embed fiddle urls as html attributes
-const fiddleUrls = () => tree => {
-  const regex = /fiddle='(.*)'/
-  visit(tree, 'code', node => {
-    if (node.lang && node.lang.includes('fiddle')) {
-      // retrieve and remove url from language definition
-      const url = node.lang.match(regex)[1]
-      node.lang = node.lang.replace(regex, '').trim()
-
-      // save url in data-fiddle-url html attribute
-      node.data = node.data || {}
-      node.data.hProperties = node.data.hProperties || {}
-      node.data.hProperties.dataFiddleUrl = url
-    }
-  })
-  return tree
-}
 
 function convertToUrlSlash(filePath) {
   return filePath.replace(/C:\\/g, '/').replace(/\\/g, '/')
